ubergo/requirements.txt
aiohttp==3.8.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
altgraph==0.17.3
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.6.0
asttokens==2.2.1
async-timeout==4.0.2
attrs==22.2.0
auto-py-to-exe==2.28.0
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==6.0.0
bottle==0.12.23
bottle-websocket==0.2.9
branca==0.6.0
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
click-plugins==1.1.1
cligj==0.7.2
colorama==0.4.6
comm==0.1.3
contourpy==1.0.7
coreapi==2.3.3
coreschema==0.0.4
coverage==7.2.7
cycler==0.11.0
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
dictionary==1.0
discord.py==2.1.0
Django==4.1.6
django-cors-headers==4.2.0
django-filter==22.1
django-grappelli==3.0.5
django-location-field==2.7.0
django-phonenumber-field==7.1.0
django-tastypie==0.14.6
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
docutils==0.20.1
drf-yasg==1.21.5
Eel==0.14.0
executing==1.2.0
fastjsonschema==2.17.1
filelock==3.12.0
Fiona==1.9.4.post1
folium==0.14.0
fonttools==4.39.4
fqdn==1.5.1
frozenlist==1.3.3
fsspec==2023.5.0
future==0.18.3
gensim==4.3.1
geographiclib==2.0
geopandas==0.14.0
geopy==2.4.0
gevent==22.10.2
gevent-websocket==0.10.1
google-api-core==2.11.0
google-api-python-client==2.85.0
google-auth==2.17.3
google-auth-httplib2==0.1.0
googleapis-common-protos==1.59.0
googlemaps==4.10.0
greenlet==2.0.2
gtfs-functions==2.0.8
httplib2==0.22.0
huggingface-hub==0.15.1
idna==3.4
inflection==0.5.1
iniconfig==2.0.0
ipykernel==6.23.1
ipython==8.13.2
ipython-genutils==0.2.0
isoduration==20.11.0
itypes==1.2.0
jedi==0.18.2
jenkspy==0.3.3
Jinja2==3.1.2
joblib==1.2.0
jsonpointer==2.3
jsonschema==4.17.3
jupyter-events==0.6.3
jupyter_client==8.2.0
jupyter_core==5.3.0
jupyter_server==2.5.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
Kivy==2.2.1
kivy-deps.angle==0.3.3
kivy-deps.glew==0.3.1
kivy-deps.sdl2==0.6.0
Kivy-Garden==0.1.5
kiwisolver==1.4.4
Markdown==3.4.1
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
mistune==2.0.5
multidict==6.0.4
nbclassic==1.0.0
nbclient==0.8.0
nbconvert==7.4.0
nbformat==5.8.0
nest-asyncio==1.5.6
nltk==3.8.1
notebook==6.5.4
notebook_shim==0.2.3
numpy==1.24.3
oauth2client==4.1.3
opencv-python==4.8.0.76
openrouteservice==2.3.3
packaging==23.1
pandas==2.0.1
pandocfilters==1.5.0
parso==0.8.3
pefile==2022.5.30
pendulum==2.1.2
phonenumbers==8.13.20
phonenumberslite==8.13.20
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.5.1
plotly==5.17.0
pluggy==1.0.0
plyer==2.1.0
pprintpp==0.4.0
prometheus-client==0.16.0
prompt-toolkit==3.0.38
protobuf==4.22.3
psutil==5.9.5
psycopg2==2.9.6
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycountry==22.3.5
pycountry-convert==0.7.2
pycparser==2.21
PyDrive==1.3.1
Pygments==2.15.1
pyinstaller==5.7.0
pyinstaller-hooks-contrib==2022.15
PyJWT==2.6.0
pyparsing==3.0.9
pypiwin32==223
pyproj==3.6.0
pyrsistent==0.19.3
pytest==7.3.1
pytest-cov==4.1.0
pytest-mock==3.10.0
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
python-json-logger==2.0.7
python-mimeparse==1.6.0
pytz==2022.7.1
pytzdata==2020.1
pywin32==306
pywin32-ctypes==0.2.0
pywinpty==2.0.10
PyYAML==6.0
pyzmq==25.0.2
regex==2023.5.5
repoze.lru==0.7
requests==2.28.2
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rsa==4.9
ruamel.yaml==0.17.24
ruamel.yaml.clib==0.2.7
scikit-learn==1.2.2
scipy==1.10.1
Send2Trash==1.8.2
shapely==2.0.1
six==1.16.0
smart-open==6.3.0
sniffio==1.3.0
soupsieve==2.4.1
sqlparse==0.4.3
stack-data==0.6.2
tenacity==8.2.3
terminado==0.17.1
threadpoolctl==3.1.0
tinycss2==1.2.1
tokenizers==0.13.3
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
transformers==4.29.2
twilio==8.8.0
typing_extensions==4.6.3
tzdata==2022.7
unicode==2.9
uri-template==1.2.0
uritemplate==4.1.1
urllib3==1.26.15
utm==0.7.0
validate-email==1.3
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.5.2
whichcraft==0.6.1
wordcloud==1.9.2
yarl==1.8.2
zope.event==4.6
zope.interface==5.5.2
